Enabling `--fedjax_experimental_disable_parallel=auto` that checks `jax.local_device_count() > 1` before running via `jax.pmap` (if true) or sequentially (if false)
Also adding documentation around `drop_remainder` behavior when running in parallel. This must be enforced because `jax.pmap` requires batches to be the same shape. In this case, it's an issue for batch size when remainder batches are not the full size but this will also be an issue for other dimensions (i.e. sequence length for sentence input).

PiperOrigin-RevId: 356515340