Introduce new `Stat` and `Metric` interfaces
Along with the new dataclasses, there are also reimplementations of all available metric functions in `fedjax.core.metrics`. The reimplementations are defined for a single example (e.g. for EMNIST, an example is an (image, label) pair; for Stack Overflow, an example is a sequence of words). These metric functions can then be vectorized to work over batches using `jax.vmap`. Defining metric functions this way allows us more flexibility in how to implement example masking internally when working with fixed batch sizes.

Co-authored-by: kho <wuke@google.com>
PiperOrigin-RevId: 366852075